taking into account the node removal cost
While calculating the modularity diff in the main cycle of __one_level() function we should also take into account the cost of the removal of a node from its current community, since this operation may itself have positive or negative effect on modularity. We corrected this, according to the original paper by Blondel et al. "Fast unfolding of communities in large networks"