Stop relabelling stable partitions
Calling `best_partition(graph, partition=stable)` where `stable`
is a stable partition has a high probability of returning a
permutation of `stable`. This can be prevented in
`community_louvain.__renumber(partition)` by first checking
that `partition` is valid i.e. has labels 0 to n, and if so not
renumbering.

An example of the current undesired behaviour:
```
>>> import community
>>> in_dict = {0:1, 1:0}
>>> out_dict = community.community_louvain.__renumber(in_dict)
>>> in_dict == out_dict
False
```